Very basic bashly.yml
Create a very basic bashly.yml

Run `./rndm --help` and look at the results.

We didn't write a single line of bash yet and got a professional looking
help message for a typical CLI.

If you ever wrote a bash script with such help message you know how
boring it is to write all the boilerplate code to achieve that. This is
the very first demonstration of how Bashly can make bash development
fun.