List 'curl' as a dependency
As our script is now using 'curl', it's a dependency, and we should make
it explicit.

When rndm users run the script in an environment where `curl` is not
available, we should give them a clear error message.

By adding a list of dependencies in our `bashly.yml`, Bashly
automatically creates a dependency checking logic for the final script.

Running rndm in an environment without `curl` would fail with a message
like this:

```
$  ./rndm
missing dependency: curl
```