Add '--max' flag to specify maximum random number
This is an example of how easy it is to create a flag that takes an
argument.

We are now allowing the user to specify the maximum random number to be
printed.

Look how Bashly automatically handles the mandatory argument the user
must pass to `--max` and we didn't need to spend our brain power with
such boilerplate logic!

There's a problem in our code though. If we don't explicitly specify the
maximum value our script doesn't work as expected.

Example:

```
$ ./rndm
0

$ ./rndm
0

$ ./rndm
0

$ ./rndm -w
Error: The maximum value must be an integer in the
[-1000000000,1000000000] interval
```

We need to code a default value for the `max_number`, and we're going to
do it in the next commit.