Bashly way to specify a default value
This approach of defining a default value is slightly better, as Bashly
automatically adds such information in the help message and now the user
is aware about the default value.

We still have a problem though: What if the user passes a non-integer
value? We should make our program smart enough to not crash when the
user calls it like in this example: `rndm --max foo`. Let's address this
in the next commit.