Validate --max argument as a positive integer
Here we're using a regular expression to validate the input as positive
integer. This makes our program stronger but our code is starting to
become unpleasant to read.

In the next commit we're going to start using functions with clear name
to make our code easier to read and maintain.