Move validation to a separate function
When we create a `src/lib/` dir and put bash files there, Bashly will
automatically get these files contents and put them in the final script.
This is a good way to keep our files simple and focused on specific
concerns.

Here we're creating the `src/lib/validations.sh` file.

Giving a clear name to an operation that involves regular expressions is
always good for a better reading. That's why we created the
`is_positive_integer` function.

We also created a `validate_positive_integer` function with the
responsibility to print an error message if the argument is not valid.

Now our main logic is a bit cleaner, with just one call to validate the
`max_number`.

This is separation of concerns: validation is one concern, generating
random numbers is another one.

In the next commit we're going to see how Bashly makes validations even
cleaner...