Bashly way to validate input
Bashly provides a clean way to add validation functions for our
arguments. It works like this:

- In `bashly.yml`, in the flag configuration we add a line like this:
`validate: function_name`.
- We create a function called `validate_function_name`, and now our
program run it before allowing the user input to pass.
- If the function prints any string to stdout, it is considered an
error. The string will be displayed on screen, as the error message.

In this commit we are using the `validate_positive_integer` for this
purpose.

Another benefit of validating input this way is that Bashly provides a
slightly better error message, with information about the validation
error. Example:

```
$ ./rndm --max foo
validation error in --max MAX:
The argument must be a positive integer. Given value: foo
```